zyn: modernize

+++ src/zyn/CombFilterBank.h
@@ -1,37 +1,38 @@
-#include "ValueSmoothingFilter.h"
-
 #pragma once
 
-namespace zyn {
+#include "ValueSmoothingFilter.h"
+
+namespace zyn
+{
 
 /**Comb Filter Bank for sympathetic Resonance*/
 class CombFilterBank
 {
 public:
 	CombFilterBank(float samplerate, float initgain);
-	void filterout(float *smp, uint16_t period);
+	void filterout(float *smp, int period);
 
-	static constexpr unsigned max_strings = 76 * 3;
-	static constexpr unsigned max_samples = 6048;
+	static constexpr int max_strings = 76 * 3;
+	static constexpr int max_samples = 6048;
 
 	float delays[max_strings];
 	float inputgain;
 	float outgain;
 	float gainbwd;
 
-	void setStrings(unsigned nr, float basefreq);
+	void setStrings(int nr, float basefreq);
 
 private:
 	float string_smps[max_strings][max_samples];
 	float basefreq;
-	unsigned strings_nr;
-	unsigned pos_writer;
+	int strings_nr;
+	int pos_writer;
 
 	/* for smoothing gain jump when using binary valued sustain pedal */
 	ValueSmoothingFilter gain_smoothing;
 
-	unsigned mem_size;
+	int mem_size;
 	float samplerate;
 };
 
-}
+} // namespace zyn
